feat: compact named theme (#63)
* feat: compact named theme

Adds CompactTheme under gitsvg/theme/themes/, registered as the 'compact' name in NAMED_THEMES; tighter spacings (~30% denser), smaller fonts, and pill offsets nudged to keep the breathing room similar to default. Adjacent: theme: font-size fields relax from int to float so the compact label can sit at 9.5; default-theme SVG output now writes font-size with a trailing .0.


* docs: trim font-size float entry


---------

+++ gitsvg/theme/_default_theme.py
@@ -189,17 +189,17 @@ class DefaultTheme(Theme):
         return "Inter, sans-serif"
 
     @classmethod
-    def _resolve_label_font_size(cls) -> int:
+    def _resolve_label_font_size(cls) -> float:
         """Font size (px) for commit-message labels."""
         return 11
 
     @classmethod
-    def _resolve_branch_label_font_size(cls) -> int:
+    def _resolve_branch_label_font_size(cls) -> float:
         """Font size (px) for branch-name pills and PR title pills."""
         return 11
 
     @classmethod
-    def _resolve_hash_font_size(cls) -> int:
+    def _resolve_hash_font_size(cls) -> float:
         """Font size (px) for the secondary hash line on commit labels."""
         return 9
